add test for metric optimality parameter
