+++ tests/test_evaluation.py
@@ -1,4 +1,4 @@
-from catwalk.evaluation import ModelEvaluator, generate_binary_at_x, fpr
+from catwalk.evaluation import ModelEvaluator, generate_binary_at_x
 import testing.postgresql
 
 import numpy
@@ -166,13 +166,3 @@ def test_generate_binary_at_x():
         [1, 1, 0, 0, 0, 0, 0, 0, 0, 0]
 
 
-def test_fpr():
-    predictions_binary = \
-        [1, 1, 1, 0, 0, 0, 0, 0]
-    labels = \
-        [1, 1, 0, 1, 0, 0, 0, 1]
-
-    result = fpr([], predictions_binary, labels, [])
-    # false positives = 1
-    # total negatives = 4
-    assert result == 0.25
